--- registercode/services/UserService.java
@@ -1,17 +1,0 @@
-package com.revature.services;
-
-import java.time.LocalDate;
-
-import com.revature.beans.User;
-import com.revature.beans.UserType;
-
-import reactor.core.publisher.Mono;
-
-public interface UserService {
-	public Mono<User> login(String username, String password);
-
-	public Mono<User> register(String username, String password, String email, String firstName, String lastName,
-			LocalDate birthday, UserType type);
-	
-	boolean checkAvailability(String newName);
-}
