+++ src/main/java/com/revature/services/UserService.java
@@ -12,4 +12,6 @@ public interface UserService {
 
 	public Mono<User> register(String username, String password, String email, String firstName, String lastName,
 			LocalDate birthday, UserType type);
+	
+	boolean checkAvailability(String newName);
 }
